requirements.txt
aiohttp
beautifulsoup4
google-genai
rich